requirement.txt
weaviate-client
sentence-transformers
groq
python-dotenv